lotresearch/src/scanner/run_scanner.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
设备扫描模块主程序

该模块负责大规模IoT设备的主动探测，支持分布式扫描和智能调度。
"""

import argparse
import logging
import yaml
from pathlib import Path

# 配置日志
logging.basicConfig(level=logging.INFO,
                    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class Scanner:
    """设备扫描器类"""
    
    def __init__(self, config):
        """初始化扫描器
        
        Args:
            config: 配置字典
        """
        self.config = config
        self.scanner_type = config.get('scanner_type', 'zmap')
        self.target = config.get('target', '127.0.0.1')
        self.ports = config.get('ports', '80,443')
        self.rate = config.get('rate', 10000)
        
    def run(self):
        """执行扫描任务"""
        logger.info(f"开始扫描，目标: {self.target}, 端口: {self.ports}")
        logger.info(f"使用扫描器: {self.scanner_type}, 速率: {self.rate} packets/s")
        
        # 这里将实现具体的扫描逻辑
        # 1. 根据配置选择扫描工具
        # 2. 执行扫描任务
        # 3. 收集扫描结果
        # 4. 存储扫描数据
        
        logger.info("扫描完成")
        return True

def load_config(config_path):
    """加载配置文件
    
    Args:
        config_path: 配置文件路径
        
    Returns:
        配置字典
    """
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config

def main():
    """主函数"""
    parser = argparse.ArgumentParser(description='IoT设备扫描器')
    parser.add_argument('--config', type=str, default='config/config.yaml',
                        help='配置文件路径')
    parser.add_argument('--target', type=str, help='扫描目标')
    parser.add_argument('--ports', type=str, help='扫描端口')
    
    args = parser.parse_args()
    
    # 加载配置
    config = load_config(args.config)
    
    # 命令行参数优先级高于配置文件
    if args.target:
        config['target'] = args.target
    if args.ports:
        config['ports'] = args.ports
    
    # 初始化并运行扫描器
    scanner = Scanner(config)
    scanner.run()

if __name__ == '__main__':
    main()
