lotresearch/requirements.txt
paho-mqtt>=1.6.1
pyyaml>=6.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
pymongo>=4.4.0
redis>=4.6.0
plotly>=5.16.0
dash>=2.10.0
pyshark>=0.5.3
